backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.10.0
python-multipart==0.0.10
python-telegram-bot==21.0.1
python-dotenv==1.0.0
gunicorn==21.2.0 